+++ src/components/contents/detail-contents.tsx
@@ -2,7 +2,7 @@ import PreviewContents from "./preview-contents";
 import {
   getDetailContents,
   getTrailersContent,
-} from "@/app/api/contents/route";
+} from "@/services/contents/route";
 import Seasons from "@/components/seasons/seasons";
 import { DetailContents, Trailers } from "@/types/contents/types";
 import DetailContentsInfo from "./detail-contents-info";
